requirements.txt
annotated-types==0.6.0
anyio==3.7.1
exceptiongroup==1.2.0
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.6
pydantic==1.10.13
PyJWT==1.7.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
alembic==1.12.0
FastAPI-SQLAlchemy==0.2.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic_core==2.14.6
SQLAlchemy==2.0.23
python-dateutil==2.8.2
redis==4.5.5
s3transfer==0.9.0
six==1.16.0
urllib3==1.26.18
python-multipart==0.0.6
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
gunicorn==21.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
rsa==4.9
json5==0.9.14
PyYAML==6.0.1
celery==5.3.6
sendgrid==6.11.0
rich==13.7.0
asgi-correlation-id[celery]==4.3.0
confluent-kafka==2.3.0